github has assets that can be attached, and "default" assets that zip or tar+gzip the contents of a release/tag
There is a need for some projects, like python projects to just pull the default assets and not have to build or attach custom
assets to a tag/release. I am adding the functionality that if you pass 'tarball' in the resource name, it will download the
tarball default asset for a release. This had to be done with "tarball_url" that comes with the release api cargo.
It also had a different accept header (application/json) as compared with the asset url

I chose overloading the asset name with "tarball" so as to keep it simple and not add more options to the resource.